docs(modal): move to ui folder
