+++ src/data/about/museum.json
@@ -11,7 +11,7 @@
   },
   {
     "id": "3",
-    "media": "museum/img1.webp",
+    "media": "museum/img3.webp",
     "description": " У 1932 році майже всіх членів товариства звинуватили у буржуазному націоналізмі. Михайла Вериківського звинуватили у надмірній творчій увазі до історичної тематики, до української народної пісні, національного колориту. Моральний дискомфорт, пов’язаний із неможливістю повної реалізації творчих задумів і широкого виконання власних творів, Михайло Вериківський відчував до кінця свого життя. Але, не дивлячись на різні утиски, М. І. Вериківський був новатором у системі пошуку новітнього музичного мислення. Ключовими тенденціями усієї європейської культури першої половини XX століття був пошук нових елементів сучасної музичної мови. Величезний вплив на становлення художньо-естетичної, творчої, виконавської, композиторської культури мали праці Б. Яворського – видатного теоретика і вченого світового масштабу, учнем якого був М. Вериківський. Опанування теорії ладового ритму Б. Яворського – досить важливий етап творчого становлення М. Вериківського. Композитор намагався не лише втілити, а й переосмислити в руслі національних традицій особливості нової технології організації музичної мови.Михайло Вериківський мріяв про те, щоб народ мав власну культуру."
   },
   {
